move RuleSnoozeBadge to the domain folder
